factor `test_urcrypt.c` into separate suite files
