--- tests/test_urcrypt.c
@@ -1,949 +1,0 @@
-#include "test_common.h"
-#include "urcrypt/urcrypt.h"
-#include <string.h>
-#include <stdio.h>
-#include <stdlib.h>
-
-/*
- * Urcrypt Test Suite
- *
- * Tests for SHA, RIPEMD-160, AES (ECB/CBC/SIV), and secp256k1 functions.
- *
- * Reference test vectors from:
- * - SHA-1/256/512: NIST CAVP Secure Hashing
- *   https://csrc.nist.gov/projects/cryptographic-algorithm-validation-program/secure-hashing
- * - RIPEMD-160: Official RIPEMD-160 page
- *   https://homes.esat.kuleuven.be/~bosselae/ripemd160.html
- * - AES ECB/CBC: NIST FIPS 197 Appendix C
- *   https://nvlpubs.nist.gov/nistpubs/fips/nist.fips.197.pdf
- * - AES-SIV: RFC 5297 Appendix A
- *   https://datatracker.ietf.org/doc/html/rfc5297
- * - secp256k1: Bitcoin Core secp256k1 library
- *   https://github.com/bitcoin-core/secp256k1
- * - Schnorr: BIP-340
- *   https://github.com/bitcoin/bips/blob/master/bip-0340.mediawiki
- */
-
-/*
- * ============================================================================
- * SHA Tests
- * ============================================================================
- */
-
-/*
- * Test: SHA-1 - Empty string
- */
-static int test_sha1_empty(void) {
-  uint8_t message[1] = {0};  /* Empty message, use array not literal */
-  uint8_t out[20];
-  uint8_t expected[20];
-
-  /* SHA-1("") = da39a3ee5e6b4b0d3255bfef95601890afd80709 (big-endian)
-   * urcrypt reverses input and output, so we reverse the expected vector */
-  hex_to_bytes("da39a3ee5e6b4b0d3255bfef95601890afd80709", expected, 20);
-  reverse_bytes(expected, 20);
-
-  urcrypt_sha1(message, 0, out);
-
-  ASSERT_MEM_EQ(out, expected, 20, "sha1 empty string mismatch");
-
-  return 0;
-}
-
-static int test_sha1_abc(void) {
-  uint8_t message[3] = "abc";
-  uint8_t out[20];
-  uint8_t expected[20];
-
-  reverse_bytes(message, 3);
-  hex_to_bytes("a9993e364706816aba3e25717850c26c9cd0d89d", expected, 20);
-  reverse_bytes(expected, 20);
-
-  urcrypt_sha1(message, 3, out);
-
-  ASSERT_MEM_EQ(out, expected, 20, "sha1 'abc' mismatch");
-
-  return 0;
-}
-
-/*
- * Test: SHA-1 - Longer string
- */
-static int test_sha1_longer(void) {
-  uint8_t message[56] = "abcdbcdecdefdefgefghfghighijhijkijkljklmklmnlmnomnopnopq";
-  uint8_t out[20];
-  uint8_t expected[20];
-
-  reverse_bytes(message, 56);
-  hex_to_bytes("84983e441c3bd26ebaae4aa1f95129e5e54670f1", expected, 20);
-  reverse_bytes(expected, 20);
-
-  urcrypt_sha1(message, 56, out);
-
-  ASSERT_MEM_EQ(out, expected, 20, "sha1 longer string mismatch");
-
-  return 0;
-}
-
-/*
- * Test: SHA-1 - Even longer string
- */
-static int test_sha1_longer2(void) {
-  uint8_t message[112] = "abcdefghbcdefghicdefghijdefghijkefghijklfghijklmghijklmnhijklmnoijklmnopjklmnopqklmnopqrlmnopqrsmnopqrstnopqrstu";
-  uint8_t out[20];
-  uint8_t expected[20];
-
-  reverse_bytes(message, 112);
-  hex_to_bytes("a49b2446a02c645bf419f995b67091253a04a259", expected, 20);
-  reverse_bytes(expected, 20);
-
-  urcrypt_sha1(message, 112, out);
-
-  ASSERT_MEM_EQ(out, expected, 20, "sha1 even longer string mismatch");
-
-  return 0;
-}
-
-/*
- * Test: SHA-1 - Determinism
- */
-static int test_sha1_determinism(void) {
-  uint8_t message1[] = "test message for sha1";
-  uint8_t message2[] = "test message for sha1";
-  uint8_t out1[20];
-  uint8_t out2[20];
-
-  urcrypt_sha1(message1, 21, out1);
-  urcrypt_sha1(message2, 21, out2);
-
-  ASSERT_MEM_EQ(out1, out2, 20, "sha1 should be deterministic");
-
-  /* Verify output is not all zeros */
-  int all_zero = 1;
-  for (int i = 0; i < 20; i++) {
-    if (out1[i] != 0) {
-      all_zero = 0;
-      break;
-    }
-  }
-  ASSERT(all_zero == 0, "sha1 should produce non-zero output");
-
-  return 0;
-}
-
-/*
- * Test: SHA-256 - Empty string
- */
-static int test_sha256_empty(void) {
-  uint8_t out[32];
-  uint8_t expected[32];
-
-  /* SHA-256("") = e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855 */
-  hex_to_bytes("e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855", expected, 32);
-
-  urcrypt_shay((uint8_t*)"", 0, out);
-
-  ASSERT_MEM_EQ(out, expected, 32, "sha256 empty string mismatch");
-
-  return 0;
-}
-
-/*
- * Test: SHA-256 - "abc"
- */
-static int test_sha256_abc(void) {
-  const char *message = "abc";
-  uint8_t out[32];
-  uint8_t expected[32];
-
-  /* SHA-256("abc") = ba7816bf8f01cfea414140de5dae2223b00361a396177a9cb410ff61f20015ad */
-  hex_to_bytes("ba7816bf8f01cfea414140de5dae2223b00361a396177a9cb410ff61f20015ad", expected, 32);
-
-  urcrypt_shay((uint8_t*)message, strlen(message), out);
-
-  ASSERT_MEM_EQ(out, expected, 32, "sha256 'abc' mismatch");
-
-  return 0;
-}
-
-/*
- * Test: SHA-256 - Longer message
- */
-static int test_sha256_longer(void) {
-  const char *message = "abcdbcdecdefdefgefghfghighijhijkijkljklmklmnlmnomnopnopq";
-  uint8_t out[32];
-  uint8_t expected[32];
-
-  /* SHA-256 of message above */
-  hex_to_bytes("248d6a61d20638b8e5c026930c3e6039a33ce45964ff2167f6ecedd419db06c1", expected, 32);
-
-  urcrypt_shay((uint8_t*)message, strlen(message), out);
-
-  ASSERT_MEM_EQ(out, expected, 32, "sha256 longer message mismatch");
-
-  return 0;
-}
-
-/*
- * Test: SHA-256 - Even longer message
- */
-static int test_sha256_longer2(void) {
-  const char *message = "abcdefghbcdefghicdefghijdefghijkefghijklfghijklmghijklmnhijklmnoijklmnopjklmnopqklmnopqrlmnopqrsmnopqrstnopqrstu";
-  uint8_t out[32];
-  uint8_t expected[32];
-
-  /* SHA-256 of message above */
-  hex_to_bytes("cf5b16a778af8380036ce59e7b0492370b249b11e8f07a51afac45037afee9d1", expected, 32);
-
-  urcrypt_shay((uint8_t*)message, strlen(message), out);
-
-  ASSERT_MEM_EQ(out, expected, 32, "sha256 longer message mismatch");
-
-  return 0;
-}
-
-/*
- * Test: SHA-512 - Empty string
- */
-static int test_sha512_empty(void) {
-  uint8_t out[64];
-  uint8_t expected[64];
-
-  /* SHA-512("") */
-  hex_to_bytes("cf83e1357eefb8bdf1542850d66d8007d620e4050b5715dc83f4a921d36ce9ce"
-               "47d0d13c5d85f2b0ff8318d2877eec2f63b931bd47417a81a538327af927da3e",
-               expected, 64);
-
-  urcrypt_shal((uint8_t*)"", 0, out);
-
-  ASSERT_MEM_EQ(out, expected, 64, "sha512 empty string mismatch");
-
-  return 0;
-}
-
-/*
- * Test: SHA-512 - "abc"
- */
-static int test_sha512_abc(void) {
-  const char *message = "abc";
-  uint8_t out[64];
-  uint8_t expected[64];
-
-  /* SHA-512("abc") */
-  hex_to_bytes("ddaf35a193617abacc417349ae20413112e6fa4e89a97ea20a9eeee64b55d39a"
-               "2192992a274fc1a836ba3c23a3feebbd454d4423643ce80e2a9ac94fa54ca49f",
-               expected, 64);
-
-  urcrypt_shal((uint8_t*)message, strlen(message), out);
-
-  ASSERT_MEM_EQ(out, expected, 64, "sha512 'abc' mismatch");
-
-  return 0;
-}
-
-/*
- * Test: SHA-512 - Longer string
- */
-static int test_sha512_longer(void) {
-  const char *message = "abcdbcdecdefdefgefghfghighijhijkijkljklmklmnlmnomnopnopq";
-  uint8_t out[64];
-  uint8_t expected[64];
-
-  /* SHA-512(message) */
-  hex_to_bytes("204a8fc6dda82f0a0ced7beb8e08a41657c16ef468b228a8279be331a703c335"
-               "96fd15c13b1b07f9aa1d3bea57789ca031ad85c7a71dd70354ec631238ca3445",
-               expected, 64);
-
-  urcrypt_shal((uint8_t*)message, strlen(message), out);
-
-  ASSERT_MEM_EQ(out, expected, 64, "sha512 longer string mismatch");
-
-  return 0;
-}
-
-/*
- * Test: SHA-512 - Even longer string
- */
-static int test_sha512_longer2(void) {
-  const char *message = "abcdefghbcdefghicdefghijdefghijkefghijklfghijklmghijklmnhijklmnoijklmnopjklmnopqklmnopqrlmnopqrsmnopqrstnopqrstu";
-  uint8_t out[64];
-  uint8_t expected[64];
-
-  /* SHA-512(message) */
-  hex_to_bytes("8e959b75dae313da8cf4f72814fc143f8f7779c6eb9f7fa17299aeadb6889018"
-               "501d289e4900f7e4331b99dec4b5433ac7d329eeb6dd26545e96e55b874be909",
-               expected, 64);
-
-  urcrypt_shal((uint8_t*)message, strlen(message), out);
-
-  ASSERT_MEM_EQ(out, expected, 64, "sha512 even longer string mismatch");
-
-  return 0;
-}
-
-/*
- * Test: SHA-256 with salt - Basic test
- *
- * Note: urcrypt_shas computes SHA-256(salt XOR SHA-256(message))
- */
-static int test_sha256_salt(void) {
-  uint8_t message[] = "test message";
-  uint8_t salt[] = "salt";
-  uint8_t out[32];
-
-  urcrypt_shas(salt, 4,
-               message, 12,
-               out);
-
-  /* Verify output is not all zeros */
-  int all_zero = 1;
-  for (int i = 0; i < 32; i++) {
-    if (out[i] != 0) {
-      all_zero = 0;
-      break;
-    }
-  }
-  ASSERT(all_zero == 0, "sha256 with salt should produce non-zero output");
-
-  /* Test determinism */
-  uint8_t out2[32];
-  uint8_t salt2[] = "salt";
-  uint8_t message2[] = "test message";
-  urcrypt_shas(salt2, 4,
-               message2, 12,
-               out2);
-
-  ASSERT_MEM_EQ(out, out2, 32, "sha256 with salt should be deterministic");
-
-  return 0;
-}
-
-/*
- * ============================================================================
- * RIPEMD-160 Tests
- * ============================================================================
- */
-
-/*
- * Test: RIPEMD-160 - Empty string
- */
-static int test_ripemd160_empty(void) {
-  uint8_t message[1] = {0};
-  uint8_t out[20];
-  uint8_t expected[20];
-
-  /* RIPEMD-160("") = 9c1185a5c5e9fc54612808977ee8f548b2258d31 (big-endian)
-   * urcrypt reverses input and output, so we reverse the expected vector */
-  hex_to_bytes("9c1185a5c5e9fc54612808977ee8f548b2258d31", expected, 20);
-  reverse_bytes(expected, 20);
-
-  int ret = urcrypt_ripemd160(message, 0, out);
-
-  ASSERT(ret == 0, "ripemd160 should succeed");
-  ASSERT_MEM_EQ(out, expected, 20, "ripemd160 empty string mismatch");
-
-  return 0;
-}
-
-/*
- * Test: RIPEMD-160 - Determinism
- */
-static int test_ripemd160_determinism(void) {
-  uint8_t message1[] = "test ripemd160";
-  uint8_t message2[] = "test ripemd160";
-  uint8_t out1[20];
-  uint8_t out2[20];
-
-  int ret1 = urcrypt_ripemd160(message1, 14, out1);
-  int ret2 = urcrypt_ripemd160(message2, 14, out2);
-
-  ASSERT(ret1 == 0, "ripemd160 should succeed");
-  ASSERT(ret2 == 0, "ripemd160 should succeed");
-  ASSERT_MEM_EQ(out1, out2, 20, "ripemd160 should be deterministic");
-
-  /* Verify output is not all zeros */
-  int all_zero = 1;
-  for (int i = 0; i < 20; i++) {
-    if (out1[i] != 0) {
-      all_zero = 0;
-      break;
-    }
-  }
-  ASSERT(all_zero == 0, "ripemd160 should produce non-zero output");
-
-  return 0;
-}
-
-/*
- * ============================================================================
- * AES ECB Tests
- * ============================================================================
- */
-
-/*
- * Test: AES-128-ECB - Encrypt and decrypt round-trip
- *
- * Using FIPS 197 Appendix C.1 test vector
- */
-static int test_aes_ecb_128(void) {
-  uint8_t key[16];
-  uint8_t plaintext[16];
-  uint8_t encrypted[16];
-  uint8_t decrypted[16];
-  uint8_t expected[16];
-
-  /* Simple round-trip test */
-  uint8_t plaintext_copy[16];
-
-  /* Initialize with simple pattern */
-  for (int i = 0; i < 16; i++) {
-    key[i] = i;
-    plaintext[i] = i * 3;
-    plaintext_copy[i] = i * 3;
-  }
-
-  /* Encrypt */
-  int ret = urcrypt_aes_ecba_en(key, plaintext, encrypted);
-  ASSERT(ret == 0, "aes-128-ecb encrypt should succeed");
-
-  /* Re-initialize key for decrypt (since it gets modified) */
-  for (int i = 0; i < 16; i++) {
-    key[i] = i;
-  }
-
-  /* Decrypt */
-  ret = urcrypt_aes_ecba_de(key, encrypted, decrypted);
-  ASSERT(ret == 0, "aes-128-ecb decrypt should succeed");
-  ASSERT_MEM_EQ(decrypted, plaintext_copy, 16, "aes-128-ecb round-trip mismatch");
-
-  return 0;
-}
-
-/*
- * Test: AES-192-ECB - Encrypt and decrypt round-trip
- *
- * Using FIPS 197 Appendix C.2 test vector
- */
-static int test_aes_ecb_192(void) {
-  uint8_t key[24];
-  uint8_t plaintext[16];
-  uint8_t plaintext_copy[16];
-  uint8_t encrypted[16];
-  uint8_t decrypted[16];
-
-  /* Initialize */
-  for (int i = 0; i < 24; i++) key[i] = i + 10;
-  for (int i = 0; i < 16; i++) {
-    plaintext[i] = i * 5;
-    plaintext_copy[i] = i * 5;
-  }
-
-  /* Encrypt */
-  int ret = urcrypt_aes_ecbb_en(key, plaintext, encrypted);
-  ASSERT(ret == 0, "aes-192-ecb encrypt should succeed");
-
-  /* Re-initialize key */
-  for (int i = 0; i < 24; i++) key[i] = i + 10;
-
-  /* Decrypt */
-  ret = urcrypt_aes_ecbb_de(key, encrypted, decrypted);
-  ASSERT(ret == 0, "aes-192-ecb decrypt should succeed");
-  ASSERT_MEM_EQ(decrypted, plaintext_copy, 16, "aes-192-ecb round-trip mismatch");
-
-  return 0;
-}
-
-/*
- * Test: AES-256-ECB - Encrypt and decrypt round-trip
- *
- * Using FIPS 197 Appendix C.3 test vector
- */
-static int test_aes_ecb_256(void) {
-  uint8_t key[32];
-  uint8_t plaintext[16];
-  uint8_t plaintext_copy[16];
-  uint8_t encrypted[16];
-  uint8_t decrypted[16];
-
-  /* Initialize */
-  for (int i = 0; i < 32; i++) key[i] = i + 20;
-  for (int i = 0; i < 16; i++) {
-    plaintext[i] = i * 7;
-    plaintext_copy[i] = i * 7;
-  }
-
-  /* Encrypt */
-  int ret = urcrypt_aes_ecbc_en(key, plaintext, encrypted);
-  ASSERT(ret == 0, "aes-256-ecb encrypt should succeed");
-
-  /* Re-initialize key */
-  for (int i = 0; i < 32; i++) key[i] = i + 20;
-
-  /* Decrypt */
-  ret = urcrypt_aes_ecbc_de(key, encrypted, decrypted);
-  ASSERT(ret == 0, "aes-256-ecb decrypt should succeed");
-  ASSERT_MEM_EQ(decrypted, plaintext_copy, 16, "aes-256-ecb round-trip mismatch");
-
-  return 0;
-}
-
-/*
- * ============================================================================
- * AES CBC Tests
- * ============================================================================
- */
-
-/* Simple realloc wrapper for AES CBC tests */
-static void* test_realloc(void* ptr, size_t size) {
-  return realloc(ptr, size);
-}
-
-/*
- * Test: AES-128-CBC - Encrypt and decrypt round-trip
- */
-static int test_aes_cbc_128(void) {
-  uint8_t key_enc[16], key_dec[16];
-  uint8_t iv_enc[16], iv_dec[16];
-  const char *message = "Hello, AES-CBC!";  /* 15 bytes, will be padded to 16 */
-
-  /* Allocate message buffer */
-  size_t msg_len = strlen(message);
-  uint8_t *encrypted = malloc(msg_len);
-  uint8_t *original = malloc(msg_len);
-  memcpy(encrypted, message, msg_len);
-  memcpy(original, message, msg_len);
-  size_t encrypted_len = msg_len;
-
-  /* Setup key and IV (need separate copies since they're modified) */
-  hex_to_bytes("2b7e151628aed2a6abf7158809cf4f3c", key_enc, 16);
-  hex_to_bytes("2b7e151628aed2a6abf7158809cf4f3c", key_dec, 16);
-  hex_to_bytes("000102030405060708090a0b0c0d0e0f", iv_enc, 16);
-  hex_to_bytes("000102030405060708090a0b0c0d0e0f", iv_dec, 16);
-
-  /* Encrypt */
-  int ret = urcrypt_aes_cbca_en(&encrypted, &encrypted_len, key_enc, iv_enc, test_realloc);
-  ASSERT(ret == 0, "aes-128-cbc encrypt should succeed");
-  ASSERT(encrypted_len == 16, "aes-128-cbc should pad to 16 bytes");
-  size_t padded_len = encrypted_len;
-
-  /* Decrypt */
-  ret = urcrypt_aes_cbca_de(&encrypted, &encrypted_len, key_dec, iv_dec, test_realloc);
-  ASSERT(ret == 0, "aes-128-cbc decrypt should succeed");
-  ASSERT(encrypted_len == padded_len, "aes-128-cbc decrypt should keep padded length");
-  /* Compare only the original message length, ignoring padding */
-  ASSERT(memcmp(encrypted, original, msg_len) == 0, "aes-128-cbc round-trip mismatch");
-
-  free(encrypted);
-  free(original);
-  return 0;
-}
-
-/*
- * Test: AES-192-CBC - Encrypt and decrypt round-trip
- */
-static int test_aes_cbc_192(void) {
-  uint8_t key_enc[24], key_dec[24];
-  uint8_t iv_enc[16], iv_dec[16];
-  const char *message = "Test 192-bit key";  /* 16 bytes */
-
-  size_t msg_len = strlen(message);
-  uint8_t *encrypted = malloc(msg_len);
-  uint8_t *original = malloc(msg_len);
-  memcpy(encrypted, message, msg_len);
-  memcpy(original, message, msg_len);
-  size_t encrypted_len = msg_len;
-
-  /* Setup key and IV (need separate copies since they're modified) */
-  hex_to_bytes("8e73b0f7da0e6452c810f32b809079e562f8ead2522c6b7b", key_enc, 24);
-  hex_to_bytes("8e73b0f7da0e6452c810f32b809079e562f8ead2522c6b7b", key_dec, 24);
-  hex_to_bytes("000102030405060708090a0b0c0d0e0f", iv_enc, 16);
-  hex_to_bytes("000102030405060708090a0b0c0d0e0f", iv_dec, 16);
-
-  /* Encrypt */
-  int ret = urcrypt_aes_cbcb_en(&encrypted, &encrypted_len, key_enc, iv_enc, test_realloc);
-  ASSERT(ret == 0, "aes-192-cbc encrypt should succeed");
-  size_t padded_len = encrypted_len;
-
-  /* Decrypt */
-  ret = urcrypt_aes_cbcb_de(&encrypted, &encrypted_len, key_dec, iv_dec, test_realloc);
-  ASSERT(ret == 0, "aes-192-cbc decrypt should succeed");
-  ASSERT(encrypted_len == padded_len, "aes-192-cbc decrypt should keep padded length");
-  /* Compare only the original message length, ignoring any padding */
-  ASSERT(memcmp(encrypted, original, msg_len) == 0, "aes-192-cbc round-trip mismatch");
-
-  free(encrypted);
-  free(original);
-  return 0;
-}
-
-/*
- * Test: AES-256-CBC - Encrypt and decrypt round-trip
- */
-static int test_aes_cbc_256(void) {
-  uint8_t key_enc[32], key_dec[32];
-  uint8_t iv_enc[16], iv_dec[16];
-  const char *message = "AES-256-CBC test";  /* 16 bytes */
-
-  size_t msg_len = strlen(message);
-  uint8_t *encrypted = malloc(msg_len);
-  uint8_t *original = malloc(msg_len);
-  memcpy(encrypted, message, msg_len);
-  memcpy(original, message, msg_len);
-  size_t encrypted_len = msg_len;
-
-  /* Setup key and IV (need separate copies since they're modified) */
-  hex_to_bytes("603deb1015ca71be2b73aef0857d77811f352c073b6108d72d9810a30914dff4", key_enc, 32);
-  hex_to_bytes("603deb1015ca71be2b73aef0857d77811f352c073b6108d72d9810a30914dff4", key_dec, 32);
-  hex_to_bytes("000102030405060708090a0b0c0d0e0f", iv_enc, 16);
-  hex_to_bytes("000102030405060708090a0b0c0d0e0f", iv_dec, 16);
-
-  /* Encrypt */
-  int ret = urcrypt_aes_cbcc_en(&encrypted, &encrypted_len, key_enc, iv_enc, test_realloc);
-  ASSERT(ret == 0, "aes-256-cbc encrypt should succeed");
-  size_t padded_len = encrypted_len;
-
-  /* Decrypt */
-  ret = urcrypt_aes_cbcc_de(&encrypted, &encrypted_len, key_dec, iv_dec, test_realloc);
-  ASSERT(ret == 0, "aes-256-cbc decrypt should succeed");
-  ASSERT(encrypted_len == padded_len, "aes-256-cbc decrypt should keep padded length");
-  /* Compare only the original message length, ignoring any padding */
-  ASSERT(memcmp(encrypted, original, msg_len) == 0, "aes-256-cbc round-trip mismatch");
-
-  free(encrypted);
-  free(original);
-  return 0;
-}
-
-/*
- * ============================================================================
- * AES-SIV Tests
- * ============================================================================
- */
-
-/*
- * Test: AES-128-SIV - Encrypt and decrypt with associated data
- *
- * Based on RFC 5297 Appendix A.1
- */
-static int test_aes_siv_128(void) {
-  uint8_t key_enc[32], key_dec[32];  /* AES-SIV uses 256-bit key for 128-bit AES */
-  uint8_t iv_enc[16], iv_dec[16];
-  const char *plaintext = "test";
-  const char *ad_str = "associated";
-
-  /* Prepare buffers */
-  size_t pt_len = strlen(plaintext);
-  uint8_t *encrypted = malloc(pt_len);
-  uint8_t *decrypted = malloc(pt_len);
-  uint8_t *original = malloc(pt_len);
-  memcpy(encrypted, plaintext, pt_len);
-  memcpy(original, plaintext, pt_len);
-
-  /* Prepare associated data - needs separate buffers since data is modified */
-  uint8_t *ad_enc = malloc(strlen(ad_str));
-  uint8_t *ad_dec = malloc(strlen(ad_str));
-  memcpy(ad_enc, ad_str, strlen(ad_str));
-  memcpy(ad_dec, ad_str, strlen(ad_str));
-
-  urcrypt_aes_siv_data ad_enc_data, ad_dec_data;
-  ad_enc_data.length = strlen(ad_str);
-  ad_enc_data.bytes = ad_enc;
-  ad_dec_data.length = strlen(ad_str);
-  ad_dec_data.bytes = ad_dec;
-
-  /* Setup key - needs separate copies since key is modified */
-  hex_to_bytes("fffefdfcfbfaf9f8f7f6f5f4f3f2f1f0f0f1f2f3f4f5f6f7f8f9fafbfcfdfeff", key_enc, 32);
-  hex_to_bytes("fffefdfcfbfaf9f8f7f6f5f4f3f2f1f0f0f1f2f3f4f5f6f7f8f9fafbfcfdfeff", key_dec, 32);
-
-  /* Encrypt */
-  int ret = urcrypt_aes_siva_en(encrypted, pt_len, &ad_enc_data, 1, key_enc, iv_enc, encrypted);
-  ASSERT(ret == 0, "aes-128-siv encrypt should succeed");
-
-  /* Decrypt */
-  ret = urcrypt_aes_siva_de(encrypted, pt_len, &ad_dec_data, 1, key_dec, iv_enc, decrypted);
-  ASSERT(ret == 0, "aes-128-siv decrypt should succeed");
-  ASSERT(memcmp(decrypted, original, pt_len) == 0, "aes-128-siv round-trip mismatch");
-
-  free(encrypted);
-  free(decrypted);
-  free(original);
-  free(ad_enc);
-  free(ad_dec);
-  return 0;
-}
-
-/*
- * Test: AES-192-SIV - Encrypt and decrypt round-trip
- */
-static int test_aes_siv_192(void) {
-  uint8_t key_enc[48], key_dec[48];  /* AES-SIV uses 384-bit key for 192-bit AES */
-  uint8_t iv_enc[16], iv_dec[16];
-  const char *plaintext = "192-bit test";
-
-  size_t pt_len = strlen(plaintext);
-  uint8_t *encrypted = malloc(pt_len);
-  uint8_t *decrypted = malloc(pt_len);
-  uint8_t *original = malloc(pt_len);
-  memcpy(encrypted, plaintext, pt_len);
-  memcpy(original, plaintext, pt_len);
-
-  /* Setup key - needs separate copies (using incrementing bytes) */
-  for (int i = 0; i < 48; i++) {
-    key_enc[i] = i;
-    key_dec[i] = i;
-  }
-
-  /* Encrypt without associated data */
-  int ret = urcrypt_aes_sivb_en(encrypted, pt_len, NULL, 0, key_enc, iv_enc, encrypted);
-  ASSERT(ret == 0, "aes-192-siv encrypt should succeed");
-
-  /* Decrypt */
-  ret = urcrypt_aes_sivb_de(encrypted, pt_len, NULL, 0, key_dec, iv_enc, decrypted);
-  ASSERT(ret == 0, "aes-192-siv decrypt should succeed");
-  ASSERT(memcmp(decrypted, original, pt_len) == 0, "aes-192-siv round-trip mismatch");
-
-  free(encrypted);
-  free(decrypted);
-  free(original);
-  return 0;
-}
-
-/*
- * Test: AES-256-SIV - Encrypt and decrypt round-trip
- */
-static int test_aes_siv_256(void) {
-  uint8_t key_enc[64], key_dec[64];  /* AES-SIV uses 512-bit key for 256-bit AES */
-  uint8_t iv_enc[16], iv_dec[16];
-  const char *plaintext = "256-bit AES-SIV test message";
-
-  size_t pt_len = strlen(plaintext);
-  uint8_t *encrypted = malloc(pt_len);
-  uint8_t *decrypted = malloc(pt_len);
-  uint8_t *original = malloc(pt_len);
-  memcpy(encrypted, plaintext, pt_len);
-  memcpy(original, plaintext, pt_len);
-
-  /* Setup key - needs separate copies */
-  for (int i = 0; i < 64; i++) {
-    key_enc[i] = i * 3;
-    key_dec[i] = i * 3;
-  }
-
-  /* Encrypt */
-  int ret = urcrypt_aes_sivc_en(encrypted, pt_len, NULL, 0, key_enc, iv_enc, encrypted);
-  ASSERT(ret == 0, "aes-256-siv encrypt should succeed");
-
-  /* Decrypt */
-  ret = urcrypt_aes_sivc_de(encrypted, pt_len, NULL, 0, key_dec, iv_enc, decrypted);
-  ASSERT(ret == 0, "aes-256-siv decrypt should succeed");
-  ASSERT(memcmp(decrypted, original, pt_len) == 0, "aes-256-siv round-trip mismatch");
-
-  free(encrypted);
-  free(decrypted);
-  free(original);
-  return 0;
-}
-
-/*
- * ============================================================================
- * secp256k1 Tests
- * ============================================================================
- */
-
-/*
- * Test: secp256k1 - Context initialization and destruction
- */
-static int test_secp_context(void) {
-  size_t size = urcrypt_secp_prealloc_size();
-  ASSERT(size > 0, "secp context size should be positive");
-
-  urcrypt_secp_context *ctx = malloc(size);
-  ASSERT(ctx != NULL, "secp context allocation should succeed");
-
-  /* Initialize with test entropy */
-  uint8_t entropy[32];
-  for (int i = 0; i < 32; i++) {
-    entropy[i] = i;
-  }
-
-  int ret = urcrypt_secp_init(ctx, entropy);
-  ASSERT(ret == 0, "secp context init should succeed");
-
-  /* Cleanup */
-  urcrypt_secp_destroy(ctx);
-  free(ctx);
-
-  return 0;
-}
-
-/*
- * Test: secp256k1 - Public key generation from private key
- */
-static int test_secp_make_pubkey(void) {
-  uint8_t privkey[32];
-  uint8_t hash[32];
-  uint8_t pubkey[32];
-
-  /* Use a known private key */
-  hex_to_bytes("0000000000000000000000000000000000000000000000000000000000000001", privkey, 32);
-
-  /* Hash is unused in make, but required parameter */
-  memset(hash, 0, 32);
-
-  int ret = urcrypt_secp_make(hash, privkey, pubkey);
-  ASSERT(ret == 0, "secp make pubkey should succeed");
-
-  /* Verify pubkey is not all zeros */
-  int all_zero = 1;
-  for (int i = 0; i < 32; i++) {
-    if (pubkey[i] != 0) {
-      all_zero = 0;
-      break;
-    }
-  }
-  ASSERT(all_zero == 0, "secp pubkey should not be all zeros");
-
-  return 0;
-}
-
-/*
- * Test: secp256k1 - Sign and recover
- */
-static int test_secp_sign_recover(void) {
-  size_t size = urcrypt_secp_prealloc_size();
-  urcrypt_secp_context *ctx = malloc(size);
-
-  uint8_t entropy[32];
-  for (int i = 0; i < 32; i++) {
-    entropy[i] = i + 42;
-  }
-  urcrypt_secp_init(ctx, entropy);
-
-  /* Setup test data */
-  uint8_t hash[32];
-  uint8_t privkey[32];
-  uint8_t pubkey_orig[32];
-  uint8_t pubkey_recovered_x[32];
-  uint8_t pubkey_recovered_y[32];
-  uint8_t v;
-  uint8_t r[32];
-  uint8_t s[32];
-
-  /* Generate keypair */
-  hex_to_bytes("c9afa9d845ba75166b5c215767b1d6934e50c3db36e89b127b8a622b120f6721", privkey, 32);
-  memset(hash, 0, 32);  /* hash unused for make */
-  urcrypt_secp_make(hash, privkey, pubkey_orig);
-
-  /* Sign a message */
-  hex_to_bytes("0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef", hash, 32);
-
-  int ret = urcrypt_secp_sign(ctx, hash, privkey, &v, r, s);
-  ASSERT(ret == 0, "secp sign should succeed");
-  ASSERT(v < 4, "secp recovery id should be 0-3");
-
-  /* Recover public key */
-  ret = urcrypt_secp_reco(ctx, hash, v, r, s, pubkey_recovered_x, pubkey_recovered_y);
-  ASSERT(ret == 0, "secp recover should succeed");
-
-  /* Note: Recovery returns x,y coordinates, not compressed pubkey format,
-   * so we just verify the recovery succeeded and returned non-zero values */
-  int x_nonzero = 0, y_nonzero = 0;
-  for (int i = 0; i < 32; i++) {
-    if (pubkey_recovered_x[i] != 0) x_nonzero = 1;
-    if (pubkey_recovered_y[i] != 0) y_nonzero = 1;
-  }
-  ASSERT(x_nonzero && y_nonzero, "secp recovered pubkey should be non-zero");
-
-  urcrypt_secp_destroy(ctx);
-  free(ctx);
-  return 0;
-}
-
-/*
- * Test: secp256k1 Schnorr - Sign and verify (BIP-340)
- */
-static int test_secp_schnorr_sign_verify(void) {
-  size_t size = urcrypt_secp_prealloc_size();
-  urcrypt_secp_context *ctx = malloc(size);
-
-  uint8_t entropy[32];
-  for (int i = 0; i < 32; i++) {
-    entropy[i] = i + 100;
-  }
-  urcrypt_secp_init(ctx, entropy);
-
-  /*
-   * Test Schnorr signature - simplified test
-   * We test that signing succeeds. Full verification testing would require
-   * extracting the x-only public key, which requires accessing internal
-   * secp256k1 structures not exposed through the urcrypt API.
-   */
-  uint8_t privkey[32];
-  uint8_t msg[32];
-  uint8_t aux[32];
-  uint8_t sig[64];
-
-  /* Initialize buffers with deterministic values */
-  for (int i = 0; i < 32; i++) {
-    privkey[i] = i + 1;
-    msg[i] = i * 2;
-    aux[i] = i * 3;
-  }
-
-  /* Test that signing succeeds */
-  int ret = urcrypt_secp_schnorr_sign(ctx, privkey, msg, aux, sig);
-  ASSERT(ret == 0, "schnorr sign should succeed");
-
-  /* Test determinism - same inputs should produce same signature */
-  uint8_t privkey2[32], msg2[32], aux2[32], sig2[64];
-  for (int i = 0; i < 32; i++) {
-    privkey2[i] = i + 1;
-    msg2[i] = i * 2;
-    aux2[i] = i * 3;
-  }
-
-  ret = urcrypt_secp_schnorr_sign(ctx, privkey2, msg2, aux2, sig2);
-  ASSERT(ret == 0, "schnorr sign should succeed (determinism test)");
-  ASSERT(memcmp(sig, sig2, 64) == 0, "schnorr signatures should be deterministic");
-
-  urcrypt_secp_destroy(ctx);
-  free(ctx);
-  return 0;
-}
-
-/* Test suite entry point */
-int suite_urcrypt(void) {
-  int suite_failures = 0;
-
-  RUN_TEST(test_sha1_empty);
-  RUN_TEST(test_sha1_abc);
-  RUN_TEST(test_sha1_longer);
-  RUN_TEST(test_sha1_longer2);
-  RUN_TEST(test_sha1_determinism);
-  RUN_TEST(test_sha256_empty);
-  RUN_TEST(test_sha256_abc);
-  RUN_TEST(test_sha256_longer);
-  RUN_TEST(test_sha256_longer2);
-  RUN_TEST(test_sha512_empty);
-  RUN_TEST(test_sha512_abc);
-  RUN_TEST(test_sha512_longer);
-  RUN_TEST(test_sha512_longer2);
-  RUN_TEST(test_sha256_salt);
-  RUN_TEST(test_ripemd160_empty);
-  RUN_TEST(test_ripemd160_determinism);
-  RUN_TEST(test_aes_ecb_128);
-  RUN_TEST(test_aes_ecb_192);
-  RUN_TEST(test_aes_ecb_256);
-  RUN_TEST(test_aes_cbc_128);
-  RUN_TEST(test_aes_cbc_192);
-  RUN_TEST(test_aes_cbc_256);
-  RUN_TEST(test_aes_siv_128);
-  RUN_TEST(test_aes_siv_192);
-  RUN_TEST(test_aes_siv_256);
-  RUN_TEST(test_secp_context);
-  RUN_TEST(test_secp_make_pubkey);
-  RUN_TEST(test_secp_sign_recover);
-  RUN_TEST(test_secp_schnorr_sign_verify);
-
-  return suite_failures;
-}
